second commit. almost there but with some problems.
Admittedly, the scale is a little off. The logits are indeed off. With reference
to the baselines on sample.jpg, we can see that 23464 strongly predicts a car
(or something) is present. When we look at the logits, we get the following
boxes.

Darknet
23460, 1.108489
23461, -0.594607
23462, -0.656502
23463, -0.234685

TF
23460 1.00243
23461 -0.738462
23462 -0.677512
23463 -0.225319

It is thus understandable that the locations would be slightly off.
Time to fine-tune the model.