completed darknet transfer. 100% identical.
Boxes and activations are identical short of numerical errors. Previous commit
suggested that we should fine-tune the model. No fine-tuning was done. Upon
closer inspection of all the logits, it was found that there were indexing
mistakes in the 13x13x425 logits when processing. When this was solved, the
activations were identical in all layers, short of some numerical errors. The
bounding boxes produced are identical (if not almost) to the original Tiny-YOLO
in Darknet.